move public tpu utils up a folder level
